rename image viewer package remove MyClass.kt files
