--- Domain/actors/src/main/java/com/sanaa/actors/MyClass.kt
@@ -1,4 +1,0 @@
-package com.sanaa.actors
-
-class MyClass {
-}
